Creating initial repository.
This is a sample application which contains many opportunities for
applying SOLID principles. Once we have a prototype working for our
project, it is a good opportunity to do a refactor to apply SOLID where
needed so that the code is more testable, extensible, readable and
maintainable in the long run.